requirement.txt
eel
python-escpos